move vest suite next to refactored form that uses it
